pyproject.toml
[tool.poetry]
name = "transformers_tutorial"
version = "0.1.0"
description = "https://huggingface.co/docs/transformers/quicktour"
authors = ["rym903 <nakagawa.ryoma.84n@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.4"
pandas = "^1.5.1"
torch = "^1.12.1"
requests = "^2.28.1"
transformers = "^4.23.1"
datasets = "^2.6.1"
Pillow = "^9.2.0"
black = "^22.10.0"
torchvision = "^0.13.1"
jupyter = "^1.0.0"
timm = "^0.6.11"
seaborn = "^0.12.1"
vit-pytorch = "^0.36.2"
tweepy = "^4.12.0"
python-dotenv = "^0.21.0"
lxml = "^4.9.1"
selenium = "^4.6.0"

[tool.poetry.dev-dependencies]
numpy = "^1.23.4"
pandas = "^1.5.1"
torch = "^1.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
